fix: switched to use the enabled properties
Each test starts with checking if it was enabled via the BOSH property.
Credhub and autoscaler tests got a refactor as they should not check for
statefulsets to determine if credhub and autoscaler are enabled on
KubeCF. This logic should be done on the CF side when requesting the
tests to run.